refactor(controller): split routes/settings.ts into mounted sub-routers
1150 lines and 20 routes covering four unrelated concerns. settings.ts is
now just the mount table (23 lines); each sub-router owns one concern and
is mounted onto the parent:

  core.ts     the settings read/write surface + credential writes
  llm.ts      provider probing and model discovery (read-only)
  tts.ts      voice preview and the voice catalogue
  station.ts  station actions: mixer, stream, themes, search probe

The paths are disjoint, so mount order carries no routing meaning — it's
ordered for readability only, and the comment says so.

Verified by enumerating the assembled router's stack at runtime and
diffing against the 20 routes the pre-split file registered: identical,
same methods, same paths.

No behaviour change. Controller lint clean, all 52 test files pass.
